Update attribute handling for booleans
When an attribute is attempted to be set to `true`, it is set as if it were
a boolean attribute, such as

```xml
<element foo />
```

This is consistent with most expectations of boolean handling.

The behaviour of false is to remove the attribute (assuming it is set).
This might be used where a pipeline is generating a series of similar
elements, and a feature is then turned off for some of them.

Finally, the documentation around the use of `null` has been made clearer:
it leaves the attribute as it currently is, perform no action.